pyproject.toml
[tool.poetry]
name = "fashion-snap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
pillow = ">=8.0,<11.0"
torch = "2.1.0"
torchvision = "0.16.0"
torchaudio = "2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
indent-width = 2

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = true




